Add files via upload
In the file deriveFR5st.nb, the polynomials P1<R1, ..., P51,..., P55 are computed in symbolic form from which the order conditions for the five\stage FR scheme of the fourth order of accuracy are obtained as described in my paper submitted to CASC 2024 Conference.  
It is to be noted that the obtained expressions for the above polynomials can be used also at the investigation of the FR scheme of lower orders of accuracy. Example: in the case of the three-stage FR scheme, one must set to zero the following parameters of the FR method: c4,d4,c5,d5, See for details my paper.

The obtaining of the Groebner bases from the order conditions iscarried out by using the Mathematica function GroebnerBasis[...]. This is NOT my function, this is a built-in Mathematica function. So far as I can see after several years of the use of this function, it contains no errors and can cope with the tasks, where the package Maple 17 fails 
(Maple informs about the shortage of memory).

The developer of the file deriveFR5st.nb: E.V. Vorozhtsov, Russia
June9, 2024